+++ NEWS.md
@@ -48,6 +48,7 @@
     "Open" dialog properly.
   * Windows rp-config: Fix the Key Manager tab not showing up at all.
     * This regressed in v2.2.1.
+  * Nintendo3DS: Added non-standard MIME types used by Citra.
 
 * Other changes:
   * Nintendo3DS: The "Options" menu no longer shows a grayed-out "Extract SRL"